chore: move accountPool to separate service
