+++ backend/src/services/transactionManager/queue.ts
@@ -2,7 +2,7 @@ import { ApiPromise } from "@polkadot/api";
 import { KeyringPair } from "@polkadot/keyring/types";
 import { Queue, TransactionStatus } from "../../models/index.js";
 import { getOnChainNonce } from "../../utils/networkApi.js";
-import { getAccount, getAccounts } from "./accountPool.js";
+import { getAccount, getAccounts } from "../accountManager/index.js";
 import { queueConfig } from "./config.js";
 import { Transaction } from "./types.js";
